+++ workspace/packages/components/siyuan/misc/index.ts
@@ -15,21 +15,20 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import type { ComponentProps } from "svelte";
+import type { TooltipsDirection } from "./tooltips";
 
-import type { TooltipsDirection } from "../../misc/tooltips";
-
-export interface IIcon {
-    icon: string; // 图标 ID
-    type?: string; // data-type 属性值
-    active?: boolean; // 是否激活
-    ariaLabel?: string; // 提示标签内容
+export interface IBlockIcon {
+    icon: string; // svg 图标引用 ID
+    none?: boolean; // 是否隐藏 .fn__none (display: none)
+    show?: boolean; // 是否显示 .block__icon--show (opacity: 1)
+    active?: boolean; // 是否激活 .toolbar__item--active
+    disabled?: boolean; // 是否禁用 .toolbar__item--disabled
+    type?: string; // data-type
+    ariaLabel?: string; // 提示标签内容 aria-label
     tooltipsDirection?: TooltipsDirection; // 提示标签方向
-    onClick?: (e: MouseEvent, element: HTMLElement, active: boolean) => boolean; // 按钮点击回调函数
-}
-
-export interface IBar {
-    logo: string; // svg 图标引用 ID
-    title: string; // 标题
-    icons: IIcon[]; // 按钮列表
+    onClick?: (
+        e: MouseEvent,
+        element: HTMLElement,
+        active: boolean,
+    ) => boolean; // 按钮点击回调函数
 }